requirements.txt
opencv-python>=4.10.0
numpy>=1.26.0
Pillow>=10.3.0
pydantic>=2.8.0
typer>=0.12.3
